@@ -5,29 +5,32 @@
   "scripts": {},
   "private": true,
   "dependencies": {
-    "@nrwl/nx-plugin": "^13.4.3",
+    "@nrwl/nx-plugin": "15.6.3",
     "tslib": "^2.0.0"
   },
   "devDependencies": {
-    "@nrwl/cli": "13.4.3",
-    "@nrwl/devkit": "13.4.3",
-    "@nrwl/eslint-plugin-nx": "13.4.3",
-    "@nrwl/jest": "13.4.3",
-    "@nrwl/js": "13.4.3",
-    "@nrwl/linter": "13.4.3",
-    "@nrwl/node": "13.4.3",
-    "@nrwl/tao": "13.4.3",
-    "@nrwl/workspace": "13.4.3",
-    "@types/jest": "27.0.2",
-    "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "~5.3.0",
-    "@typescript-eslint/parser": "~5.3.0",
-    "eslint": "8.2.0",
+    "@nrwl/cli": "15.6.3",
+    "@nrwl/devkit": "15.6.3",
+    "@nrwl/eslint-plugin-nx": "15.6.3",
+    "@nrwl/jest": "15.6.3",
+    "@nrwl/js": "15.6.3",
+    "@nrwl/linter": "15.6.3",
+    "@nrwl/node": "15.6.3",
+    "@nrwl/workspace": "15.6.3",
+    "@swc-node/register": "^1.4.2",
+    "@swc/core": "^1.2.173",
+    "@types/jest": "28.1.8",
+    "@types/node": "18.7.1",
+    "@typescript-eslint/eslint-plugin": "5.51.0",
+    "@typescript-eslint/parser": "5.51.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
-    "jest": "27.2.3",
-    "prettier": "^2.3.1",
-    "ts-jest": "27.0.5",
+    "jest": "28.1.3",
+    "prettier": "2.8.4",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
     "tslib": "^2.0.0",
-    "typescript": "~4.4.3"
+    "typescript": "4.8.4",
+    "nx": "15.6.3"
   }
 }
